+++ python/cudf/cudf/tests/test_orc.py
@@ -205,10 +205,12 @@ def test_orc_read_filtered(datadir, engine):
     # Create dataframes to compare
     df = cudf.read_orc(path, engine=engine)
     df_filtered = cudf.read_orc(path, engine=engine, filters=[[("int1", "==", 1)]])
+    df_filtered_le = cudf.read_orc(path, engine=engine, filters=[[("int1", "<=", 2)]])
 
     # Compare
     assert len(df) == 11000
     assert len(df_filtered) == 5000
+    assert len(df_filtered_le) == 10000
 
 
 @pytest.mark.parametrize("engine", ["cudf", "pyarrow"])